tests/wpt/web-platform-tests/app-history/navigate/return-value/reload-rejection-order-unload-unserializablestate.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>

<iframe id="i" src="/common/blank.html"></iframe>

<script>
promise_test(async t => {
  await new Promise(resolve => window.onload = resolve);

  let navigateEventCount = 0;
  i.contentWindow.appHistory.onnavigate = () => navigateEventCount++;
  i.contentWindow.appHistory.onnavigatesuccess = t.unreached_func("onnavigatesuccess should not be called");
  i.contentWindow.appHistory.onnavigateerror = t.unreached_func("onnavigateerror should not be called");

  i.contentWindow.appHistory.navigate("?1");

  await new Promise(resolve => {
    i.contentWindow.onunload = t.step_func(async () => {
      await assertBothRejectDOM(t, i.contentWindow.appHistory.reload({ state: document.body }), "InvalidStateError", i.contentWindow);
      assert_equals(navigateEventCount, 1);
      resolve();
    });
  });
}, `reload() with an unserializable state inside onunload throws "InvalidStateError", not "DataCloneError"`);
</script>
